moving check_random_state() to skimage._shared.util
